rename pins.h
