+++ Marlin/src/pins/pins.h
@@ -712,7 +712,7 @@
 #elif MB(MKS_SKIPR_V1)
   #include "stm32f4/pins_MKS_SKIPR_V1_0.h"      // STM32F4                                env:mks_skipr_v1 env:mks_skipr_v1_nobootloader
 #elif MB(TRONXY_V10)
-  #include "stm32f4/pins_TRONXY_F446.h"         // STM32F4                                env:STM32F446_tronxy
+  #include "stm32f4/pins_TRONXY_V10.h"         // STM32F4                                env:STM32F446_tronxy
 
 //
 // ARM Cortex M7
